chore: Stop being an esm module
Instead we'll use .mjs files. Storybook didn't like it when
its main file was in a type: module repo
