+++ .github/workflows/percy-components-o-viewport.yml
@@ -32,7 +32,7 @@ jobs:
         cache: 'npm'
         node-version: 16
     - run: npm ci
-    - run: node ./scripts/percy.js
+    - run: node ./scripts/percy.mjs
       env:
         PERCY_TOKEN: ${{ secrets.PERCY_TOKEN_COMPONENTS_O_VIEWPORT }}
         WORKSPACE: components/o-viewport
