--- .github/workflows/sync-repo-labels.yml
@@ -1,12 +1,0 @@
-name: Sync Origami labels
-on:
-  - issues
-  - pull_request_target
-jobs:
-  sync-labels:
-    runs-on: ubuntu-latest
-    name: Sync repository labels
-    steps:
-      - uses: Financial-Times/origami-labels@v1
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
