+++ package.json
@@ -3,11 +3,9 @@
 	"name": "origami",
 	"version": "1.0.0",
 	"description": "the reference implementation of the FT's design system",
-	"type": "module",
 	"scripts": {
-		"regenerate": "node ./scripts/regenerate-ci-config-files.js",
-		"test": "node ./scripts/test.js",
-		"install": "cp hooks/* .git/hooks && cd tools/origami-build-tools && npm ci"
+		"regenerate": "node ./scripts/regenerate-ci-config-files.mjs",
+    "test": "node ./scripts/test.mjs",
 	},
 	"repository": {
 		"type": "git",
