+++ scripts/test.mjs
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import {$} from "zx"
-import * as workspaces from "./lib/workspaces.js"
+import * as workspaces from "./lib/workspaces.mjs"
 import {readPackage} from "read-pkg"
 
 let testFilter = process.env.ORIGAMI_TEST_FILTER
