+++ templates/percy-workflow.yml
@@ -33,7 +33,7 @@ jobs:
         cache: 'npm'
         node-version: 16
     - run: npm ci
-    - run: node ./scripts/percy.js
+    - run: node ./scripts/percy.mjs
       env:
         PERCY_TOKEN: ${{ secrets.PERCY_TOKEN_<% percyTokenName %> }}
         WORKSPACE: <%& workspace %>
